UPSTREAM:DP:wifi:ath12k: Update mu_group_id, user_pos, ppdu_id and ru_start to ensure accurate stats accounting
1. mu_group_id in tx_stats not being updated, causing incorrect MU
   group info in tx_stats in current implementation.
   This change retrieve group_id and user_pos from
   HTT_PPDU_STATS_TAG_USR_RATE TLV and update tx_stats fields to ensure
   accurate MU group reporting.

2. In ppdu_id(32 bits) FW metadata is present in upper 8 bits which is
   unique for each PPDU. Current implementation uses the entire 32 bits
   to get ppdu_desc which is incorrect.
   This change mask ppdu_id to use only lower 24 bits to maintain list
   integrity

3. Update ru_start value in tx_stats. Current implementation always sets
   ru_start to zero. Retrieve correct value from user_rate TLV and
   update for accurate stats.

Patch-dependency: Ib70322e80b83338866cfcb8cb7e3bcc9515660ab (wifi:
ath12k: Move HTT specific code to newly introduced files)

+++ backports-6.1-9a0dddfb3/drivers/net/wireless/ath/ath12k/dp_cmn.h
@@ -67,8 +67,8 @@ struct ath12k_per_peer_tx_stats {
 	u16 ru_tones;
 	u8 ba_fails;
 	u8 ppdu_type;
-	u32 mu_grpid;
-	u32 mu_pos;
+	u8 mu_grpid;
+	u8 mu_pos;
 	bool is_ampdu;
 };
